# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Payroll",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employee-deductions": {
            "post": {
                "tags": [
                    "Employee Deductions"
                ],
                "summary": "Store employee deductions",
                "description": "Store single or multiple employee deductions. For bulk operations, use the 'deductions' array.",
                "operationId": "bca11d1399d8c89ff349779f87804409",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "required": [
                                            "payroll_period_id",
                                            "employee_id",
                                            "deduction_id"
                                        ],
                                        "properties": {
                                            "payroll_period_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "employee_id": {
                                                "type": "integer",
                                                "example": 123
                                            },
                                            "deduction_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "billing_cycle": {
                                                "type": "string",
                                                "example": "monthly"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1000
                                            },
                                            "percentage": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10.5
                                            },
                                            "date_from": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-01-01"
                                            },
                                            "date_to": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-12-31"
                                            },
                                            "with_terms": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "total_term": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "total_paid": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "is_default": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "isDifferential": {
                                                "type": "string",
                                                "example": "no"
                                            },
                                            "reason": {
                                                "type": "string",
                                                "example": "Monthly loan payment"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            },
                                            "deduct_at": {
                                                "type": "string",
                                                "example": "yes"
                                            },
                                            "stopped_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null
                                            },
                                            "completed_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "required": [
                                            "payroll_period_id",
                                            "deductions"
                                        ],
                                        "properties": {
                                            "payroll_period_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "deductions": {
                                                "type": "array",
                                                "items": {
                                                    "required": [
                                                        "employee_number",
                                                        "deduction_id"
                                                    ],
                                                    "properties": {
                                                        "employee_number": {
                                                            "type": "string",
                                                            "example": "EMP001"
                                                        },
                                                        "deduction_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "amount": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 1000
                                                        },
                                                        "total_term": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "total_paid": {
                                                            "type": "integer",
                                                            "example": 0
                                                        },
                                                        "notes": {
                                                            "type": "string",
                                                            "example": "Monthly deduction"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Deductions saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data successfully saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or employee not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Employee EMP001 not found."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "employee_id": [
                                                    "The employee id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}